requirements.txt
pandas==1.1.4
matplotlib==3.3.3
seaborn==0.11.1
numpy==1.18.5
streamlit==0.72.0
plotly==4.14.1
Pillow==8.0.1
scikit_learn==0.24.0
missingno
